requirements.txt
joblib==1.4.2
numpy==1.26.4
streamlit==1.39.0
altair==5.4.1
jinja2==3.1.4
MarkupSafe==2.1.3
jsonschema==4.23.0
attrs==24.2.0
jsonschema-specifications==2023.7.1
referencing==0.30.2
rpds-py==0.10.6
narwhals==1.5.5
blinker==1.6.2
cachetools==5.5.0
click==8.1.7
colorama==0.4.6
gitpython==3.1.43
gitdb==4.0.11
smmap==5.0.1
packaging==24.1
pandas==24.1
python-dateutil==2.9.0.post0
six==1.16.0
pytz==2024.1
tzdata==2023.3
pillow==10.4.0
protobuf==4.25.5
pydeck==0.9.1
requests==2.32.3
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
urllib3==1.26.19
rich==13.7.1
markdown-it-py==2.2.0
mdurl==0.1.0
pygments==2.15.1
tenacity==8.2.3
toml==0.10.2
tornado==6.4.1
typing-extensions==4.11.0
watchdog==4.0.1
sqlalchemy==2.0.34
greenlet==3.0.1
langchain-openai==0.2.2
langchain-core==0.3.12
jsonpatch==1.33
jsonpointer==2.1
langsmith==0.1.136
httpx==0.27.0
anyio==4.6.2
exceptiongroup==1.2.0
httpcore==1.0.2
h11==0.14.0
sniffio==1.3.0
orjson==3.10.7
requests-toolbelt==1.0.0
pydantic==2.9.2
annotated-types==0.6.0
pydantic-core==2.34.4
PyYAML==6.0.2
openai==1.43.0
distro==1.9.0
jiter==0.5.0
tqdm==4.66.5
tiktoken==0.7.0
regex==2024.9.11
networkx==3.2.1
scikit-learn==1.5.2
scipy==1.31.1
threadpoolctl==3.5.0
flask==3.0.3
importlib-metadata==7.0.1
zipp==3.20.2
itsdangerous==2.2.0
Werkzeug==3.0.3
mysql-connector-python==9.0.0
requests==2.32.3
plotly==5.24.1
networkx==3.2.1
matplotlib==3.9.2
contourpy==1.3.0
cycler==0.11.0
fonttools==4.51.0
importlib-resources==6.4.0
kiwisolver==1.4.4
pyparsing==3.1.2
